% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestforge-package.R
\docType{package}
\name{backtestforge-package}
\alias{backtestforge}
\alias{backtestforge-package}
\title{backtestforge: Easy backtesting tools}
\description{
Fast native trend-following backtesting tools for stocks, futures, and crypto, with legacy blotter/quantstrat compatibility available for regression checks.
}
\author{
\strong{Maintainer}: Hugo Rzepian Teixeira \email{hugorteixeira@senhormercado.com.br}

}
\keyword{internal}
