% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBacktestTools-package.R
\docType{package}
\name{rBacktestTools-package}
\alias{rBacktestTools}
\alias{rBacktestTools-package}
\title{rBacktestTools: What the Package Does (Title Case)}
\description{
R library for making it easier to do simple backtests using blotter/quantstrat with trend following strategies on stocks, futures and crypto.
}
\author{
\strong{Maintainer}: Hugo Rzepian Teixeira \email{hugorteixeira@senhormercado.com.br}

}
\keyword{internal}
