% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestforge-package.R
\docType{package}
\name{backtestforge-package}
\alias{backtestforge}
\alias{backtestforge-package}
\title{backtestforge: Easy backtesting tools}
\description{
R library for making it easier to do simple backtests using blotter/quantstrat with trend following strategies on stocks, futures and crypto.
}
\author{
\strong{Maintainer}: Hugo Rzepian Teixeira \email{hugorteixeira@senhormercado.com.br}

}
\keyword{internal}
